Portal now has a cooldown of 8 minutes (Previously 20 minutes), but it only teleports the Imp, not nearby hunters.
Reasoning behind the change : Imp was supposed to act as a massive pressure for the runner(s) and with such a long cooldown, it felt like it didn't bring enough utility to the team over the game. Imp isn't a team player (At least in it's kit) and thus, removing the ability to move the entire team makes sense.