Deleo changes to it's main combo ability.
The combo now deals much less knockback and it's gap-closer covers less distance to compensate.

Additionally, Deleo doesn't deal damage as it breaks a shield.

(On a quick note, this commit also has changes to all characters with "shield-breaking" abilities. The shield-breaking is now more "stable"

And Berserker's shield disable simply didn't work correctly. It somebody was using a shield, it wouldnt break it; it would only prevent them from using if try to re-hold it if that makes sense.